refactor: page 생성 함수 분리
